Resume computers which are suspended
This will suspend a computer which has timed out and only resume
it again when it is ticked.

This means computers which are out of tick radius will not shut down
but instead buffer events until resumed. This is hidden behind a config
option which is disabled for obvious reasons.

Fixes #136