Rewrite some of the pocket upgrade system
Wireless modems now have a special upgrade object. This still uses an upgrade
id of 0 and is not registered for crafting. However we elusively use upgrade methods now,
and so avoid some special handling for them.

This means things like the PocketAPIExtensions class does not need special support for modems
too.